Updated handling for updated registers type. Decimal keyboard type for iOS/iPadOS/visionOS.

+++ Robotic Complex Workspace/Functions.swift
@@ -42,7 +42,7 @@ func select_tool_modules(name: String, model_controller: inout ToolModelControll
 //MARK: - Changer modules
 let changer_modules_names = ["Module", "Module 2"]
 
-func change_by(name: String, registers: inout [Int])
+func change_by(name: String, registers: inout [Float])
 {
     switch name
     {
